Add test for CheckPubKeyEncoding
Summary:
As per title. More test is better.

Depends on D1565

Test Plan:
  make check

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox, teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1566

+++ src/test/txvalidationcache_tests.cpp
@@ -107,7 +107,7 @@ void ValidateCheckInputsForAllFlags(const CMutableTransaction &mutableTx,
     PrecomputedTransactionData txdata(tx);
     // If we add many more flags, this loop can get too expensive, but we can
     // rewrite in the future to randomly pick a set of flags to evaluate.
-    for (size_t test_flags = 0; test_flags < (1U << 17); test_flags += 1) {
+    for (uint32_t test_flags = 0; test_flags < (1U << 17); test_flags += 1) {
         CValidationState state;
         // Make sure the mandatory flags are enabled.
         test_flags |= MANDATORY_SCRIPT_VERIFY_FLAGS;
